Mobility database

Database of all mobility models stored in nomad

model_db

Format

A named list of nomad_models.

Each model is named with a standard convention. For example, let's look at zmb_cdr_2020_mod_dd_exp. Each of the names in the snake case name convention refer either to the data that the model was fit using or the type of model that was fit.

zmb_cdr_2020_mod_dd_exp:

zmb: ISO3C county code for where data was collected

cdr: Type of mobility data. cdr = Call Data Record

2020: Year of data collection

mod: mod signifies the following labels relate to the model

dd: Which mobility model. dd = Departure-Diffusion

exp: Sub-Type of mobility model. exp = Exponential

The naming conventions help with documenting the models and enable linking to mobility_db to query further information about the underlying mobility data the model was fit using. All names in the snake case model name before mod refer to the mobility data.